# backends/fastapi/openapi.yaml
openapi: 3.1.0
info:
  title: Todo App API - FastAPI
  description: 'Todo 애플리케이션 REST API.


    ## Categories

    카테고리 CRUD 엔드포인트


    ## Tasks

    카테고리별 필터링 지원 Task CRUD 엔드포인트'
  version: 1.0.0
paths:
  /api/categories:
    get:
      tags:
      - Categories
      summary: 카테고리 목록 조회
      description: 등록된 모든 카테고리를 반환함.
      operationId: get_categories_api_categories_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CategoryResponse'
                type: array
                title: Response Get Categories Api Categories Get
    post:
      tags:
      - Categories
      summary: 카테고리 생성
      description: 새 카테고리를 생성하고 생성된 리소스를 반환함.
      operationId: create_category_api_categories_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
        required: true
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/categories/{id}:
    delete:
      tags:
      - Categories
      summary: 카테고리 삭제
      description: ID로 카테고리를 삭제함. 연관된 Task도 함께 삭제됨(cascade).
      operationId: delete_category_api_categories__id__delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '204':
          description: 삭제 성공
        '404':
          description: 카테고리 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/tasks:
    get:
      tags:
      - Tasks
      summary: Task 목록 조회
      description: 전체 Task 목록을 반환함. `categoryId` 쿼리 파라미터로 카테고리 필터링 가능.
      operationId: get_tasks_api_tasks_get
      parameters:
      - name: categoryId
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: 필터링할 카테고리 ID (Optional)
          title: Categoryid
        description: 필터링할 카테고리 ID (Optional)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
                title: Response Get Tasks Api Tasks Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Tasks
      summary: Task 생성
      description: 새 Task를 생성하고 생성된 리소스를 반환함. `categoryId`가 존재하지 않으면 404 반환.
      operationId: create_task_api_tasks_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: 카테고리 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/tasks/{id}:
    put:
      tags:
      - Tasks
      summary: Task 수정
      description: ID로 Task를 조회하여 전체 업데이트(PUT semantics)함.
      operationId: update_task_api_tasks__id__put
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Task 또는 카테고리 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Tasks
      summary: Task 삭제
      description: ID로 Task를 삭제함.
      operationId: delete_task_api_tasks__id__delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '204':
          description: 삭제 성공
        '404':
          description: Task 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    CategoryRequest:
      properties:
        name:
          type: string
          title: Name
          description: 카테고리 이름
          examples:
          - Work
        color:
          type: string
          title: Color
          description: 카테고리 색상 Hex 코드
          examples:
          - '#FF5733'
      type: object
      required:
      - name
      - color
      title: CategoryRequest
    CategoryResponse:
      properties:
        id:
          type: integer
          title: Id
          description: 카테고리 ID
        name:
          type: string
          title: Name
          description: 카테고리 이름
        color:
          type: string
          title: Color
          description: 카테고리 색상 Hex 코드
      type: object
      required:
      - id
      - name
      - color
      title: CategoryResponse
    ErrorResponse:
      properties:
        detail:
          type: string
          title: Detail
          description: 에러 메시지
          examples:
          - Category not found
      type: object
      required:
      - detail
      title: ErrorResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    TaskRequest:
      properties:
        title:
          type: string
          title: Title
          description: Task 제목
          examples:
          - Submit Report
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          description: Task 상세 설명
        completed:
          type: boolean
          title: Completed
          description: 완료 여부
          default: false
        dueDate:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Duedate
          description: 마감일 (ISO-8601)
          examples:
          - '2026-05-31T23:59:59+09:00'
        priority:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          title: Priority
          description: 우선순위
          examples:
          - HIGH
        categoryId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Categoryid
          description: 연결할 카테고리 ID
      type: object
      required:
      - title
      - priority
      title: TaskRequest
    TaskResponse:
      properties:
        id:
          type: integer
          title: Id
          description: Task ID
        title:
          type: string
          title: Title
          description: Task 제목
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          description: Task 상세 설명
        completed:
          type: boolean
          title: Completed
          description: 완료 여부
        dueDate:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Duedate
          description: 마감일 (ISO-8601)
        priority:
          type: string
          title: Priority
          description: 우선순위 (LOW, MEDIUM, HIGH)
        category:
          anyOf:
          - $ref: '#/components/schemas/CategoryResponse'
          - type: 'null'
          description: 연결된 카테고리
      type: object
      required:
      - id
      - title
      - completed
      - priority
      title: TaskResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError